+++ ui/src/app/modules/plugins/plugin-card/plugin-info/plugin-info.component.ts
@@ -3,6 +3,8 @@ import { Component, inject, Input, OnInit } from '@angular/core'
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap'
 import { TranslatePipe } from '@ngx-translate/core'
 
+import { Plugin } from '@/app/core/manage-plugins/manage-plugins.interfaces'
+
 @Component({
   templateUrl: './plugin-info.component.html',
   styleUrls: ['./plugin-info.component.scss'],
@@ -12,7 +14,7 @@ import { TranslatePipe } from '@ngx-translate/core'
 export class PluginInfoComponent implements OnInit {
   private $activeModal = inject(NgbActiveModal)
 
-  @Input() plugin: any
+  @Input() plugin: Plugin
 
   public readonly defaultIcon = 'assets/hb-icon.png'
   public readonly linkScoped = '<a href="https://github.com/homebridge/plugins/wiki/Scoped-Plugins" target="_blank"><i class="fas fa-fw fa-external-link-alt primary-text"></i></a>'