+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - disable plugin notes now based on keep orphans setting
 - text clarification in users support modal
 - remove glibc checks as node 18 is now unsupported
+- add restart child bridges option to plugin log modal
 
 ### Other Changes